backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    { name = "Son Pham-Ba", email = "son.phamba@epfl.ch" },
    { name = "Yannick Marcon", email = "yannick.marcon@epfl.ch" },
    { name = "Pierre Guilbert", email = "pierre.guilbert@epfl.ch" },
    { name = "Mati Shah", email = "mati.shah@epfl.ch" },
]
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "fastapi>=0.115.14",
    "uvicorn>=0.35.0",
    "networkx>=3.5",
    "numpy>=2.3.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "scipy>=1.16.0",
    "python-multipart>=0.0.20",
    "pandas>=2.3.1",
    "openpyxl>=3.1.5",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.2",
    "fastapi-cache2>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "dotenv>=0.9.9",
    "pytest>=8.4.1",
    "python-dotenv[cli]>=1.1.1",
]
test = [
    "pytest>=8.4.1",
]

[tool.codespell]
skip = 'uv.lock'
